requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.3
django-allauth==65.3.1
django-modeltranslation==0.19.12
environs==11.2.1
idna==3.10
marshmallow==3.23.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
psycopg==3.2.3
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
urllib3==2.3.0

